requirements.txt
# Instant Tutor AI - Production Requirements
# Core dependencies for AI-powered tutoring system

# Web Framework
fastapi>=0.100.0
uvicorn>=0.20.0
websockets>=12.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
redis>=5.0.0

# Vector Databases
pinecone-client>=2.2.0
weaviate-client>=3.25.0

# AI and ML
openai>=1.0.0
anthropic>=0.8.0
sentence-transformers>=2.2.0
transformers>=4.30.0
whisper>=1.1.0
SpeechRecognition>=3.10.0

# Video Processing
opencv-python>=4.8.0
moviepy>=1.0.0
ffmpeg-python>=0.2.0
librosa>=0.10.0
Pillow>=10.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.65.0
yt-dlp>=2023.12.0

# Development
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
black>=23.0.0
flake8>=6.0.0

# Security
PyJWT>=2.8.0
cryptography>=41.0.0